Refactor ChessBoard class to only handle the operations on the board
That is, initialise a board with the given size, add/remove/move a piece.
To make the chessboard more general, it should not include any game rule.
All the game rule will be on the concrete Piece subclasses such as Pawn or Rook.
In ChessBoard pass $_piece into Array $_pieces, change getName() function in Pawn, Rook.